+++ HISTORY.rst
@@ -23,6 +23,7 @@ Unreleased
 * IMP: Add the module web_sheet_full_width
 
 * ALCN-909 Rename Delivery Rounds models and load new delivery rounds setup
+* ALCN-913: Add the unique supplier on the product form
 
 **Bugfixes**